Use rate.Limit instead of untyped floats in ratelimiters
A rather minor but wide-touching change, but it will be useful for a later PR.

+++ tools/cli/admin_task_list_commands.go
@@ -116,7 +116,7 @@ func printTaskListStatus(taskListStatus *types.TaskListStatus) {
 		ReadLevel: taskListStatus.GetReadLevel(),
 		AckLevel:  taskListStatus.GetAckLevel(),
 		Backlog:   taskListStatus.GetBacklogCountHint(),
-		RPS:       taskListStatus.GetRatePerSecond(),
+		RPS:       float64(taskListStatus.GetRatePerSecond()),
 		StartID:   taskListStatus.GetTaskIDBlock().GetStartID(),
 		EndID:     taskListStatus.GetTaskIDBlock().GetEndID(),
 	}}
